Moving of Caffe loaders into separate file
