+++ modules/dnn/src/caffe/layer_loaders.hpp
@@ -39,9 +39,9 @@
 //
 //M*/
 
-#ifndef __OPENCV_DNN_LAYERS_FULLY_CONNECTED_LAYER_HPP__
-#define __OPENCV_DNN_LAYERS_FULLY_CONNECTED_LAYER_HPP__
-#include "../precomp.hpp"
+#ifndef __OPENCV_DNN_CAFFE_LAYER_LOADERS_HPP__
+#define __OPENCV_DNN_CAFFE_LAYER_LOADERS_HPP__
+
 #include <opencv2/dnn/all_layers.hpp>
 
 namespace cv
@@ -49,22 +49,9 @@ namespace cv
 namespace dnn
 {
 
-class FullyConnectedLayerImpl : public InnerProductLayer
-{
-    int axisCan, dtype;
-    int numOutput, innerSize, outerSize;
-    bool bias, useOpenCL;
-    Blob biasOnesBlob;
-
-    template<typename XMat>
-    void forward_(std::vector<Blob*> &input, std::vector<Blob> &output);
-
-public:
-
-    FullyConnectedLayerImpl(int axisCan = 1);
-    void allocate(const std::vector<Blob*> &input, std::vector<Blob> &output);
-    void forward(std::vector<Blob*> &inputs, std::vector<Blob> &outputs);
-};
+//Common template for Caffe layer loaders
+template <typename PublicLayer>
+Ptr<Layer> createLayerFromCaffe(LayerParams&);
 
 }
 }
