+++ modules/dnn/src/layers/softmax_layer.hpp
@@ -67,8 +67,6 @@ public:
     void forward(std::vector<Blob*> &inputs, std::vector<Blob> &outputs);
 };
 
-Ptr<Layer> createSoftmaxLayerFromCaffe(LayerParams &params);
-
 }
 }
 #endif
